backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
google-generativeai
supabase
pymupdf
whisperx
torch
torchaudio
torchvision
pandas
setuptools-rust
